+++ LookupAnything/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Updated for multi-key bindings in SMAPI 3.9.
+* Fixed support for mods which add gift tastes by context tag.
 * Fixed coffee crops not showing next harvest date.
 
 ## 1.32.2